move cluster to action packages and remove wip-test
