+++ pkg/actions/cluster/cluster.go
@@ -1,6 +1,7 @@
 package cluster
 
 import (
+	"github.com/kris-nova/logger"
 	api "github.com/weaveworks/eksctl/pkg/apis/eksctl.io/v1alpha5"
 	"github.com/weaveworks/eksctl/pkg/cfn/manager"
 	"github.com/weaveworks/eksctl/pkg/eks"
@@ -18,8 +19,10 @@ func New(cfg *api.ClusterConfig, ctl *eks.ClusterProvider) (Cluster, error) {
 	}
 
 	if manager.IsClusterStack(stacks) {
+		logger.Debug("Cluster %q was created by eksctl", cfg.Metadata.Name)
 		return NewOwnedCluster(cfg, ctl, stackManager)
 	}
+	logger.Debug("Cluster %q was not created by eksctl", cfg.Metadata.Name)
 
 	return NewUnownedCluster(cfg, ctl)
 }
